+++ include/game_field_area.hpp
@@ -9,28 +9,28 @@ template <class cell_t>
 struct IGameFieldArea {
     IGameFieldArea(
             std::shared_ptr<game_field::GameField<cell_t>> field,
-            std::pair<size_t, size_t> upperLeftCorner, 
-            std::pair<size_t, size_t> lowerRightCorner):
+            std::pair<int, int> upperLeftCorner, 
+            std::pair<int, int> lowerRightCorner):
         field_(field)
         , upperLeftCorner_(upperLeftCorner)
         , lowerRightCorner_(lowerRightCorner)
     {}
 
-    virtual bool isCellAvailable(size_t xidx, size_t yidx) const noexcept = 0;
-    virtual void setCell(size_t xidx, size_t yidx, const cell_t& cell) = 0;
-    virtual cell_t getCell(size_t xidx, size_t yidx) const = 0;
+    virtual bool isCellAvailable(int xidx, int yidx) const noexcept = 0;
+    virtual void setCell(int xidx, int yidx, const cell_t& cell) = 0;
+    virtual cell_t getCell(int xidx, int yidx) const = 0;
 
     virtual ~IGameFieldArea() {} 
 
 protected:
     std::shared_ptr<game_field::IGameField<cell_t>> field_;
-    std::pair<size_t> upperLeftCorner_;
-    std::pair<size_t> lowerRightCorner_;
+    std::pair<int> upperLeftCorner_;
+    std::pair<int> lowerRightCorner_;
 };
 
 template <class cell_t>
 bool IGameFieldArea<cell_t>::isCellAvailable(
-    size_t xidx, size_t yidx) const noexcept 
+    int xidx, int yidx) const noexcept 
 {
     return xidx >= upperLeftCorner_.first && 
            xidx <= lowerRightCorner_.first &&
@@ -46,28 +46,28 @@ class GameFieldArea :
 public:
     GameFieldArea(
             std::shared_ptr<game_field::GameField<cell_t>> field,
-            std::pair<size_t, size_t> upperLeftCorner, 
-            std::pair<size_t, size_t> lowerRightCorner) :
+            std::pair<int, int> upperLeftCorner, 
+            std::pair<int, int> lowerRightCorner) :
         IGameFieldArea(field, upperLeftCorner, lowerRightCorner)
     {}
 
 public:
-    bool isCellAvailable(size_t xidx, size_t yidx) const noexcept override {
+    bool isCellAvailable(int xidx, int yidx) const noexcept override {
         retrun IGameFieldArea::isCellAvailable(xidx, yidx);
     }
 
-    void setCell(size_t xidx, size_t yidx, const cell_t& cell) override {
+    void setCell(int xidx, int yidx, const cell_t& cell) override {
         verifyThenThrowCellPos_(xidx, yidx);
         field_->setCell(xidx, yidx, cell);
     }
 
-    cell_t getCell(size_t xidx, size_t yidx) const override {
+    cell_t getCell(int xidx, int yidx) const override {
         verifyThenThrowCellPos_(xidx, yidx);
         return field_->getCell(xidx, yidx);
     }
 
 private:
-    void verifyThenThrowCellPos_(size_t xidx, size_t yidx) {
+    void verifyThenThrowCellPos_(int xidx, int yidx) {
         if (isExludedCell(xidx, yidx)) {
             throw std::logic_error("Accessing a forbidden cell.");
         }
